rec_service/requirements.txt
# Core dependencies
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
pydantic-settings==2.3.4

# ML and Vector Search
sentence-transformers==3.0.1
faiss-cpu==1.8.0
numpy==1.26.4
torch==2.3.1

# Kafka
aiokafka==0.11.0
kafka-python==2.0.2

# Redis for queue management
redis==5.0.7
aioredis==2.0.1

# HTTP clients
httpx==0.27.0
aiohttp==3.9.5

# Data processing
pandas==2.2.2
pyarrow==16.1.0

# Rate limiting and retry
tenacity==8.4.2
ratelimit==2.2.1

# Monitoring and metrics
prometheus-client==0.20.0

# Testing
pytest==8.2.2
pytest-asyncio==0.23.7
pytest-cov==5.0.0
pytest-timeout==2.3.1
pytest-mock==3.14.0
faker==25.8.0

# Utilities
python-multipart==0.0.9
python-dotenv==1.0.1
orjson==3.10.5